refactor(hangar-daemon): extract acp_session::ensure from the create RPC
`fleet/acp_session_create` was the only place the `fleet_session` +
`fleet_acp_session` pair could be minted, with the registry check, the
permission-mode read, the paired insert and the live-scope replay guard
all inline in the handler. The task executor needs the exact same
transaction for `scope_key = task:<id>`, and a second copy would drift.

The body moves to `acp_session::ensure(pool, events, provider, cwd,
scope_key) -> FleetAcpSessionRow` with a typed `EnsureError`; the RPC
keeps its capability gate and parameter checks and maps the error onto
the same `invalid_params` / `internal` codes and messages as before, so
the existing socket tests pass unchanged.
